epic switch to unsigned string functions, and de-warningification

+++ libs/graphics/lcd_arraybuffer.c
@@ -16,21 +16,21 @@
 #include "jsvar.h"
 
 // returns the BIT index, so the bottom 3 bits specify the bit in the byte
-int lcdGetPixelIndex_ArrayBuffer(JsGraphics *gfx, int x, int y, int pixelCount) {
+size_t lcdGetPixelIndex_ArrayBuffer(JsGraphics *gfx, int x, int y, int pixelCount) {
   if (gfx->data.flags & JSGRAPHICSFLAGS_ARRAYBUFFER_ZIGZAG) {
     if (y&1) x = gfx->data.width - (x+pixelCount);
   }
   if (gfx->data.flags & JSGRAPHICSFLAGS_ARRAYBUFFER_VERTICAL_BYTE)
-    return ((x + (y>>3)*gfx->data.width)<<3) | (y&7);
+    return (size_t)(((x + (y>>3)*gfx->data.width)<<3) | (y&7));
   else
-    return (x + y*gfx->data.width)*gfx->data.bpp;
+    return (size_t)((x + y*gfx->data.width)*gfx->data.bpp);
 }
 
 unsigned int lcdGetPixel_ArrayBuffer(JsGraphics *gfx, short x, short y) {
   unsigned int col = 0;
   JsVar *buf = jsvObjectGetChild(gfx->graphicsVar, "buffer", 0);
   if (buf && jsvIsArrayBuffer(buf)) {
-    int idx = lcdGetPixelIndex_ArrayBuffer(gfx,x,y,1);
+    size_t idx = lcdGetPixelIndex_ArrayBuffer(gfx,x,y,1);
     JsvArrayBufferIterator it;
     jsvArrayBufferIteratorNew(&it, buf, idx>>3 );
     if (gfx->data.bpp < 8) {
@@ -54,7 +54,7 @@ unsigned int lcdGetPixel_ArrayBuffer(JsGraphics *gfx, short x, short y) {
 void lcdSetPixels_ArrayBuffer(JsGraphics *gfx, short x, short y, short pixelCount, unsigned int col) {
   JsVar *buf = jsvObjectGetChild(gfx->graphicsVar, "buffer", 0);
   if (buf && jsvIsArrayBuffer(buf)) {
-    int idx = lcdGetPixelIndex_ArrayBuffer(gfx,x,y,pixelCount);
+    size_t idx = lcdGetPixelIndex_ArrayBuffer(gfx,x,y,pixelCount);
     JsvArrayBufferIterator it;
     jsvArrayBufferIteratorNew(&it, buf, idx>>3 );
     short p;
