+++ libs/network/http/jswrap_http.c
@@ -295,24 +295,24 @@ JsVar *jswrap_url_parse(JsVar *url, bool parseQuery) {
   int addrEnd = (portStart>=0) ? portStart : pathStart;
   // pull out details
   jsvUnLock(jsvObjectSetChild(obj, "method", jsvNewFromString("GET")));
-  jsvUnLock(jsvObjectSetChild(obj, "host", jsvNewFromStringVar(url, addrStart+1, addrEnd-(addrStart+1))));
+  jsvUnLock(jsvObjectSetChild(obj, "host", jsvNewFromStringVar(url, (size_t)(addrStart+1), (size_t)(addrEnd-(addrStart+1)))));
 
   JsVar *v;
 
-  v = jsvNewFromStringVar(url, pathStart, JSVAPPENDSTRINGVAR_MAXLENGTH);
+  v = jsvNewFromStringVar(url, (size_t)pathStart, JSVAPPENDSTRINGVAR_MAXLENGTH);
   if (jsvGetStringLength(v)==0) jsvAppendString(v, "/");
   jsvUnLock(jsvObjectSetChild(obj, "path", v));
 
-  v = jsvNewFromStringVar(url, pathStart, (searchStart>=0)?(searchStart-pathStart):JSVAPPENDSTRINGVAR_MAXLENGTH);
+  v = jsvNewFromStringVar(url, (size_t)pathStart, (size_t)((searchStart>=0)?(searchStart-pathStart):JSVAPPENDSTRINGVAR_MAXLENGTH));
   if (jsvGetStringLength(v)==0) jsvAppendString(v, "/");
   jsvUnLock(jsvObjectSetChild(obj, "pathname", v));
 
-  jsvUnLock(jsvObjectSetChild(obj, "search", (searchStart>=0)?jsvNewFromStringVar(url, searchStart, JSVAPPENDSTRINGVAR_MAXLENGTH):jsvNewNull()));
+  jsvUnLock(jsvObjectSetChild(obj, "search", (searchStart>=0)?jsvNewFromStringVar(url, (size_t)searchStart, JSVAPPENDSTRINGVAR_MAXLENGTH):jsvNewNull()));
 
   if (portNumber<=0 || portNumber>65535) portNumber=80;
   jsvUnLock(jsvObjectSetChild(obj, "port", jsvNewFromInteger(portNumber)));
 
-  JsVar *query = (searchStart>=0)?jsvNewFromStringVar(url, searchStart+1, JSVAPPENDSTRINGVAR_MAXLENGTH):jsvNewNull();
+  JsVar *query = (searchStart>=0)?jsvNewFromStringVar(url, (size_t)(searchStart+1), JSVAPPENDSTRINGVAR_MAXLENGTH):jsvNewNull();
   if (parseQuery && !jsvIsNull(query)) {
     JsVar *queryStr = query;
     jsvStringIteratorNew(&it, query, 0);
